Thread-specific GraphQL schema
